requirements.txt
absl-py==1.4.0
accelerate==0.21.0
ace_tools==0.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.2.0
annotated-types==0.6.0
ansicon==1.89.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.2.0
appdirs==1.4.4
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autogluon==1.0.0
autogluon.common==1.0.0
autogluon.core==1.0.0
autogluon.features==1.0.0
autogluon.multimodal==1.0.0
autogluon.tabular==1.0.0
autogluon.timeseries==1.0.0
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==6.1.0
blessed==1.20.0
blinker==1.6.2
blis==0.7.11
boto3==1.34.23
botocore==1.34.23
branca==0.7.2
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.10
catboost==1.2.2
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.6.4
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
colorful==0.5.6
comm==0.1.4
confection==0.1.4
contourpy==1.0.7
contractions==0.1.73
coverage==7.1.0
crcmod==1.7
croniter==1.4.1
cryptography==41.0.7
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.8
dask==2023.5.0
datasets==2.10.1
dateutils==0.6.12
debugpy==1.6.7.post1
decorator==5.1.1
deepdiff==6.7.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.8
docopt==0.6.2
editor==1.6.5
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.1.0
executing==1.2.0
fake-useragent==1.5.1
fastai==2.7.13
fastapi==0.112.0
fastcore==1.5.29
fastdownload==0.0.7
fastdtw==0.3.4
fastjsonschema==2.20.0
fastprogress==1.0.3
filelock==3.13.1
fiona==1.9.6
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
folium==0.16.0
fonttools==4.39.0
frozenlist==1.4.0
fsspec==2023.12.2
funcy==2.0
future==0.18.3
gast==0.4.0
gdown==5.0.0
gensim==4.3.2
geopandas==0.13.2
gitdb==4.0.11
GitPython==3.1.42
gluonts==0.14.3
google-api-core==2.15.0
google-api-python-client==2.153.0
google-auth==2.21.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
gpustat==1.1.1
graphviz==0.20.1
greenlet==3.1.1
grpcio==1.56.0
gspread==6.0.2
h11==0.14.0
h5py==3.9.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.20.2
hyperopt==0.2.7
idna==3.4
imageio==2.33.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
inquirer==3.2.1
insta-scrape==2.1.2
ipykernel==6.25.2
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jinxed==1.2.1
jmespath==0.10.0
joblib==1.3.1
JPype1==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.17.3
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab_pygments==0.3.0
keras==2.13.1
kiwisolver==1.4.4
konlpy==0.6.0
langchain==0.2.17
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langcodes==3.3.0
langsmith==0.1.143
lazy_loader==0.3
libclang==16.0.0
lightgbm==4.1.0
lightning==2.0.9.post0
lightning-cloud==0.5.61
lightning-utilities==0.10.1
llvmlite==0.41.1
locket==1.0.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mlforecast==0.10.0
model-index==0.1.11
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.14
munpy==0.0.0
murmurhash==1.0.10
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.7
networkx==3.1
nlpaug==1.1.11
nltk==3.8.1
nptyping==2.4.1
numba==0.58.1
numexpr==2.8.6
numpy==1.24.3
nvidia-ml-py==12.535.133
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.2.3
openai==0.27.8
opencensus==0.11.4
opencensus-context==0.1.3
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.0.10
openxlab==0.0.34
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.11
oss2==2.17.0
outcome==1.2.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.3
partd==1.4.1
patsy==0.5.3
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.13.0
plotly-express==0.4.0
pluggy==1.0.0
pprintpp==0.4.0
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
py-spy==0.3.14
py4j==0.10.9.7
pyahocorasick==2.0.0
pyarrow==15.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycountry-convert==0.7.2
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.1b0
pyee==11.1.0
Pygments==2.16.1
PyJWT==2.6.0
pyLDAvis==3.4.0
pyOpenSSL==24.1.0
pyparsing==3.0.9
pyppeteer==2.0.0
pyproj==3.5.0
pyquery==2.0.0
pyreadstat==1.2.7
pyrsistent==0.20.0
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.2.1
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-twitter==3.5
pytorch-lightning==2.0.9.post0
pytorch-metric-learning==1.7.3
pytrends==4.9.0
pytz==2023.3.post1
pyupbit==0.2.33
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
ray==2.9.1
readchar==4.0.5
regex==2023.12.25
repoze.lru==0.7
requests==2.28.2
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
rich==13.4.2
rsa==4.9
Rtree==1.2.0
runs==1.2.0
s3transfer==0.10.0
safetensors==0.4.1
scikit-image==0.20.0
scikit-learn==1.3.0
scipy==1.9.1
seaborn==0.13.2
selenium==4.11.2
sentencepiece==0.1.99
seqeval==1.2.2
serverless-wsgi==3.0.5
shapely==2.0.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
stack-data==0.6.2
starlette==0.37.2
starsessions==1.3.0
statsforecast==1.4.0
statsmodels==0.13.5
streamlit==1.31.1
streamlit_folium==0.20.0
StrEnum==0.4.15
sympy==1.12
tabulate==0.9.0
tenacity==8.2.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
text-unidecode==1.3
textsearch==0.0.24
thinc==8.2.2
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==0.9.12
tinycss2==1.4.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchmetrics==1.1.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
trio==0.22.0
trio-websocket==0.10.3
tweepy==4.14.0
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.13
utilsforecast==0.0.10
uvicorn==0.26.0
validators==0.22.0
virtualenv==20.21.0
w3lib==2.2.1
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.6
weasel==0.3.4
webdriver-manager==4.0.0
webencodings==0.5.1
websocket-client==1.7.0
websockets==10.4
Werkzeug==2.3.6
window-ops==0.0.14
wordcloud==1.9.3
wrapt==1.15.0
wsproto==1.2.0
xgboost==2.0.3
xmod==1.8.1
xxhash==3.4.1
xyzservices==2024.4.0
yarg==0.1.9
yarl==1.9.2
yolk3k==0.9
zipp==3.15.0
